requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2023.3
xarray>=2023.8.0
netCDF4>=1.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
jupyter>=1.0.0
notebook>=7.0.0
